Rename file: debug.[ch] --> debug_uart.[ch]
