+++ src/CMakeLists.txt
@@ -242,8 +242,8 @@ set(bp5_common
         ui/ui_term.h
         ui/ui_statusbar.c
         ui/ui_statusbar.h
-        debug.c
-        debug.h
+        debug_uart.c
+        debug_uart.h
 
         # JSON parser for saved settings
         mjson/mjson.h
